cleaning up some warnings and types

+++ external/oscpack/zeroconf/NetServiceThread.cpp
@@ -35,8 +35,8 @@ bool NetServiceThread::poll(DNSServiceRef dnsServiceRef, double timeOutInSeconds
 	FD_SET(dns_sd_fd, &readfds);
 	
   struct timeval tv;
-  tv.tv_sec = long(floor(timeOutInSeconds));
-  tv.tv_usec = long(1000000*(timeOutInSeconds - tv.tv_sec));
+  tv.tv_sec = (floor(timeOutInSeconds));
+  tv.tv_usec = (1000000*(timeOutInSeconds - tv.tv_sec));
 	
 	int result = select(nfds,&readfds,NULL,NULL,&tv);
 	if(result>0 && FD_ISSET(dns_sd_fd, &readfds))
